HREPTWO-3403: delay service registration until after construction
The main change is that starting clusters is now discouraged during
construction.  Since registered services are not visible outside the plugin
itself, the started cluster cannot access them.

The start() and stop() methods have been moved to the IPlugin interface.
RenderPlugin or ListViewPlugin subclasses should implement the onStart() and/or
onStop() methods.  These verify that the base class implementations are called
as well.


svn path=/hippo-ecm/trunk/; revision=20769