HREPTWO-3380: parse and validate html
Use the Neko HTML parser to detect characters and images.

For HTML properties (type "Html"), validation is triggered when the "non-empty" 
validator is present.  The same validation can be added to other String
properties by adding the "html" validator.

There is also a refactoring of the i18n of messages.  The Violation now 
contains message keys and parameters, suitable for StringResourceModels.
Translations can be added to ValidationFeedback class.


svn path=/hippo-ecm/trunk/; revision=20827