HREPTWO-3468: add cascading validation
Fields are always validated, unless their type specifies that validation does not cascade by default.  This can be overridden by the "required" validator.  The cascading behavior for a type is specified in hipposysedit:cascadevalidation, which is interpreted as "true" when absent.

Also took the liberty of making the cnd serialization deal with pseudo types and moved the responsibility of coming up with a default value for a property to the property value provider.


svn path=/hippo-ecm/trunk/; revision=20883