HREPTWO-3485, HREPTWO-2995: refactor editor
Separate interfaces for template and prototype stores.  A draft jcr store will retrieve draft types from a namespace, to create a view on the types with one namespace in a transitional state.  The field names for builtin types are now also in line with the ones generated when editing types.  The template engine only uses locators for types and templates, and an interface for prototypes.  A separate template engine factory is created when editing types, using the draft jcr store.

svn path=/hippo-ecm/trunk/; revision=20958