HREPTWO-3806: do not use the plugin context to retrieve the yui (header) manager
The yui manager was made available as a service in the plugin framework, creating a tight coupling between the rich client behaviors with the plugin framework.  The behaviors now only use a wicket behavior that is attached to the Page, thereby reducing the coupling.

There is still coupling left through the configuration; most settings for the behaviors are initialized from plugin config directly.  They still need to be adapted to use the PluginConfigMapper.


svn path=/hippo-ecm/trunk/; revision=21674