HREPTWO-4120: Use new key format to resolve resource bundle for workflow impl
The format "<primary key>,<name>=<value>[,<name>=<value>[,...]]" is used to provide a set of criteria to look up translations.  The staged localizer reduces the criteria until a match is found.  The resource provider consumer adds the criteria to the criteria map sent to the provider.  The class from key loader resolves a resource bundle, even when the class itself is not available.

svn path=/hippo-ecm/trunk/; revision=22496