HREPTWO-4489: create 'provider' project, containing the interfaces for virtual providers
The repository engine no longer depends on the modules, only this new provider
project.  All tests that need virtual providers have been moved to the
repository-test project, as it's not possible to use the manifest in the tests
for the modules project itself.

The RemoteTest did not clean up after itself, this has been added to get other
tests to be able to run succesfully.

Projects that depend on the repository-engine for running tests, now need to
include a dependency to the modules, to get the virtual providers.  WAR
overlays that embed the repository-application don't need to do this.


svn path=/hippo-ecm/trunk/; revision=23845