HREPTWO-4464: unformat plugin messes up autosave
Since javascript is case-sensitive, the plugin cannot find it's toolbar item.
Fixing this seems to fix the saving, though it's not clear how it affected the
auto-save exactly.

Also including a synchronous post of the autosave when invoked in a pre-call
handler, as the POST seemed to be handled after the concurrent GET.



svn path=/hippo-ecm/trunk/; revision=23887