HREPTWO-4517: first stab at a data model
In this model, translations all share an 'id' in a
'hippotranslation:translated' mixin type.  This type provides a child node that
is populated by a virtual provider with the translations of the document.

The locale only needs to be set at the top-level folder.  Descendant folders
will inherit the locale automatically.

The folder workflow implementation has been modified to deal gracefully with
autocreated child nodes when copying.  It is stricter on required parameters,
and will throw an exception when a parameter is needed but has not been
provided.  A new magic key 'uuid' does not need to be specified; it defaults to
a random UUID.


svn path=/hippo-ecm/trunk/; revision=23926