HREPTWO-4517: use depth-first traversal in sanity check
Virtual nodes are no longer valid after a refresh(false).  So don't cache them
in the breadth-first visitor, as we're calling the workflow during the
traversal.  (and therefore need to refresh the session)

The maximum depth has been set to 10; since the visitor is certainly allowed to
enter virtual trees there does have to be some limit.


svn path=/hippo-ecm/trunk/; revision=23940