HREPTWO-4603: allow creation of translated root folders
There are two options in the root folder creation dialog, new folder and new translated folder.
The first one allows one to create untranslated folders; a translated folder can then be added at a lower level.

The folder type for the translated folder has been changed to 'new-translated-folder'.

The country icon in the editor tab has been restored, as a trivial change with translations back in the cms.

svn path=/hippo-ecm/cms/trunk/; revision=24884