HREPTWO-4819: remove precondition check for workflow menu items
Since the workflow menu is regularly rebuild (not just repainted), with new
wicket ids generated each time, the precondition check would fail.

The text component in the action description has become a label instead of a
link, as the workflow(s) plugin will make sure that a link is created.