CMS7-5191 Author is not able to translate a document
Seems like this error is happening when the document is being translated with Google with the user "author". This is probably happening because the two workflow actions are called for the document to create the "translation" and to call google translate to change the document's content. The AccessDeniedException is thrown in the Translate document which tries to save the document with userSession.
 
 As discussed with Berry, the TranslateWorkflowImpl now uses the rootSession instead of userSession to update the document.