CMS7-5317: limit the number of fields
Use case: in the CMS a user can add unlimited fields to a multiple. The site
part may only support a certain number of those fields, so the possiblilty to
set a limit to the multiple is wanted. 

To configure this, set the 'maxitems' property of a field plugin.  When 0
(zero), there is no limit.