CMS7-3953: consolidate test structure
A PluginSuite has been introduced that makes it easy to run tests in one module
against a single repository instance.  The base PluginTest has been put into
the hippo-cms-test jar, which as a test dependency will pull in everything
that's needed to create a plugin test with embedded repository.

The remodel workflow tests have been moved into the repository test, these do
still use the old pattern as a (really) clean repository is indispensible in
that case.