HSTTWO-1774: let the channel manager create properly encoded channel IDs itself based on the channel's name.
The channel IDs are no longer provided by the user when persisting a channel, but are generated from the channel's name. A channel is therefore now obliged to have a name. The generated channel IDs are properly encoded to ensure that they are also valid JCR node names. If the generated channel ID already exists, a suffix '-1' is added (or '-2'. '-3' etc. if the name + suffix also already exists). The persist() method returns the generated channel ID.