HSTTWO-1599: moved rendering host to the HTTP session.
Whenever a rendering hosts is supplied as a request parameter we add it to the HTTP session. The rendering host in the HTTP session is then also used for all subsequent requests. This fixes a lot of cases in the template composer where the rendering host parameter was not supplied due to the site's logic (e.g. when adding public request parameters for pagination or search queries). The HstLink tag only sets the rendering host for cross-domain links, as they may go to a different host. Kudos to Jannis for peer-programming this one.