CMS7-5497: deprecating & disabling the cluster sso key
The shared key is only necessary when site and cms can be on
different hosts.  In the current setup, both should always be running on the
same cluster node.  I.e. the site is accessible on the cms domain, with a
sticky session configuration on the reverse proxy for this domain.