CMS7-5564: Notification window to indicate the state of the HST configuration
The notification window gets aligned to the top toolbar of the template composer and shown/hidden with enabling of the UI. Instead of removing all buttons from the toolbar with disabling the UI they get deactivated and later replaced with activation.  

Some error handling was giving technical information on some points which get logged as errors to the console now. 

Licensing headers were missing in some js files.