CMS7-5566: open hst config editor from the channel editor
The hst config editor is now available via a menu in the channel editor. The menu is visible when the channel is edited. The PageEditor still needs to be aware of the the channelId and the hstMountPoint also when it is called via the ChannelManagerService. Probably the channelStore has to be available from the PageEditor also for the channel properties panel later on.