CMS7-5122: cache perspective and editor dom on the client
All tabbed panels are now rendered, but only one of them is shown.
The CardView in the TabbedPanel takes care of this; it created container
elements using javascript, that are then populated with regular wicket
rendering.

Markup for subclasses of the TabsPlugin now does need an additional element to
contain this cardview.

The (xinha) editor manager now has an EditorContext per document editor.
Before, it was assuming that there was only one (editor) form available on the
page.  The context now contains all xinha editors for one document editor.

Since the pluginconfiguration by default has a null name, different render
services could end up with the same markupId.  To prevent this, markupIds are
now generated using the session, like plain wicket does.  When a name /is/
provided, it will be used, leading to a stable id.