CMS7-5606: determine whether user can add channels and edit the hst configuration
Depending on whether the user has access to modify channels, show the 'add
channel' button and the 'edit' button in the template composer.
The grid view shows the 'add channel' based on a new method on the channel
manager interface, the template composer users the CMS-User header to the
composermode rest resource.

The channel manager perspective is now shown for all users, but authors have
only limited privileges.