CMS7-5617: components toolbar
The components toolbar consists of a window which is rendering a custom GridView. The window is aligned to the top of the channel editor.

The width of the toolbar still needs to be adjusted according to the amount of the items in the toolbar (have a look at the TODO).