CMS7-5686: dynamically adjust the width of the component toolbar to the number of items
Firefox 8 does not change the width of the component toolbar itself (like Chrome), so we do it in javascript. This fixes the scrolling issue in Firefox 8. Also reset the position of the toolbar every time its loaded to avoid that components of another channel are shown in a part of the toolbar that has been moved off screen.
 
Also did tweaked the UI a bit:
- toolbar is now as wide as the iframe body
- scroll buttons are now of the same height as the toolbar