CMS7-16: remove dead code
Adding the behaviors to the page does not make sense in an ajax response.  It should be done when the page is constructed, but there is no generic way (yet) for plugins to register behaviors for that purpose.  Since the channel manager perspective is now loaded and rendered on the initial page load, the javascript files are currently rendered correctly in the head.