CMS7-5712: always include the content root path of a blueprint in the ExtJS store
Even when a blueprint does not define its own content tree, the blueprint can still contain a 'site' node that points to the default content root path for that blueprint. That default content root path should then be used as the default value of the folder picker in the 'new channel' dialog.