CMS7-5802: make services and cluster-level properties available to (workflow) plugins too
All properties that are defined on a cluster level are now automatically
available too.  The inheriting plugin configuration is now explicitly populated
with the properties of the fallback configuration.

In the abstract workflow plugin, the wicket.model service is explicitly set for
the plugin to prevent the plugin from overriding the model for the workflowS
plugin.