CMS7-5875: enable adding field editors for inherited fields
The fields that are inherited from super types (having an editor configured)
can now be added to a sub-type.  These are shown in the accordion alongside the
property, node and mixin types.

A small bug-fix is included for the moving of child nodes of the prototype when
it is committed.  This was introduced in earlier type inheritance improvements.