CMS7-6243 Xinha doesn't load after deletion of field from multiple
A memory leek due to multiple context and editor registration was causing a xinha editor to be registered multiple times if a compound containing xinha got removed. The cleanup hooks to hippoajax (modal dialog callbacks) did not take action, the manager was holding all instances in memory. Cleanup is now done gradually by new registrations, if the related dom element is still present.