HSTTWO-2235: add generic http session listener that dispatches http events as application events to all modules
The HstSessionEventPublisher needs to be registered in the web.xml; it will
dispatch events to the HstSessionEventListener.  That, in turn, will wrap the
event in a spring ApplicationEvent (subclass).  The spring event is then
broadcast to all application contexts (including submodules).

The threadlocal in the application context ensures that events do not get
trapped in recursion.  Each event is only dispatched once in every context.