HSTTWO-2325 Implemented : 1) do a complete re-implementation of the event processing 2) do finegrained reloading on some *major* nodes instead of finegrained reloads in some structure 3) do not rely on property change, node add or node remove but just mark a path as changed 4) do a pre-validate in the hstManagerImpl before going to build the enhanced hst model : During pre-validate, I can already identify problems. In case of problems, a fullblown reload is done 5) introduce a ModelLoadingException that is a runtime exception that can be thrown when it is detected that the model is incorrect : When this exception is thrown, a fullblown reload is done. Note that I try to detect whether the model is loaded incorrectly by comparing with loaded sites in a previous run. This is because of course the hst configuration can also be *really* broken in jcr
Result:
Where I used to get an error within 10 seconds when publishing and editing hst configs in a cluster and at the same time request the sites, I now did not get these failures any more. 
From time to time, you see this message:
01.11.2012 14:54:30 WARN  [org.hippoecm.hst.configuration.model.HstManagerImpl.getVirtualHosts():263] Model was possibly not build correctly. A total rebuild will be done now after flushing all caches. Reason : org.hippoecm.hst.configuration.model.HstManagerImpl$ModelLoadingException: Found HST SITE '/hst:hst/hst:sites/hippogogreen' that does not have a configuration node. However, it was loaded correctly before. 

This is an indication that a retry will be done. I think it can only happen when editing and publishing a single site a couple of times within seconds (the other cluster node has pulled in some changes and send out jcr events, and then pull in more changes during reloading the model but cannot yet process the jcr events in the hst manager)

TODO I still need to check
 1) REST plain pipeline mounts that have a mountpoint that does not point to a hst:site node but for example directly to assets
  2) Whether preview mounts still work correctly with login