CMS7-6597 Fire a DOM event when a Tab is selected
The dom event was fired on document.body, with an id of the perspective, which is now improved by fireing the event on the root element of the perspective. The perspective element can be found by searching a parent node with the classname 'perspective' present. Where listeners can register, so they don't need to be aware of the id of the perspective, only the perspective they are in.