CMS7-6634: in IE, encode the URL of the composermode Ajax call.
The URL is the document.location.href of the iframe, which IE (contrary to FF and Chrome) stores unencoded. When the iframe contains the URL to a document that has special unicode characters in it, the Ajax call will fail in IE because the HST gets the wrong URL, cannot find the content bean, and the HST component then returns a 404.