CMS7-6460: use jslint and fix the errors it reports
We now suppress the warnings of the yuicompressor, and use jslint instead (which reports much more). From now on, the build fails when jslint encounters any errors. We omit the miframe code from testing since it is not ours.

Also removed the generation of -min.js files by the yuicompressor, which were bundled in the jar but never used.