CMS7-6508: improve management of path in URL, added global behavior registration to the root plugin
Moved the control of the path in the URL to PathInUrlController, which is registered by the ControllerPlugin. The root plugin now automatically registers plugins that implement IBehavior and register themselves under the name IBehavior.class.getName(). The PathInUrlController uses this mechanism to load a Javascript file once with the function to alter the path in the URL. Subsequent Ajax calls only include a call to that function in the response.