HSTTWO-2475: only redirect on GET and POST requests in the cms security valve
Other methods are effectively mapped to a GET by the redirects.  Since the request then no longer has the original method, it is not mapped to the correct (rest) resource (java) method.

Resolution is to not do the redirects, but let the client know immediately that there's no session available.