REPO-669: don't replace access manager on already instantiated session
The local ISM could end up with a different access manager as the rest of the session.
Internal FacetRule has been renamed to QFacetRule, to indicate that it deals with resolved names instead of the (unresolved) names in the api class.
Use a AuthorizationFilterPrincipal to get the domain rule extensions into the access manager.