CMS7-7307: added field plugin for the 'Formatted text' type, i.e. a normal string property containing HTML.
There are now two CKEditor plugins: the CKEditorNodePlugin for editing HTML in the hippostd:content property 
of a child node (which also supports picking internal links and images) and the CKEditorPropertyPlugin for 
editing HTML in a normal String property (without supporting internal links and images).
 
The Wicket behaviors for the internal link picker and image picker have been abstracted into a generic
behavior of the CKEditorPanel. Custom server-side behavior can now easily be added by overriding 
CKEditorNodePlugin#createEditPanel and adding more behaviors to the CKEditorPanel returned by the superclass.