REPO-707 several smaller improvements / fixes, most important fix below in fac nav engin. It was like below // this method populates the facetValueCountMap for the current facet Object[] keyObjects = {matchingDocsSetBuilder,propertyName,parsedFacet};
matchingDocsSetBuilder was used as 'key' but does not have a equals or hashcode. Instead, use the toBitSet as key as this one has a natural and cheap hashcode & equals. 

Also renamed MultiReaderQueryFilder to CachingMultiReaderQueryFilder as this seemed more appropriate to me. 

Finally, now allow for SetDocIdSetBuilder to have a add(null) : Don't really like it myself yet, however, made code in fac nav engine simpler.

Also changed if (++x % size) into a check with == as this did result in considerable speedup