CMS7-7357: let CreateTypeDialog validate errors before executing the action
The Wicket 6 implementation for retrieving the errors in the feedback panel now also 
uses the Page of the feedback panel. Hence we need the dialog to still be attached to 
a Page when the errors are validated. The action removes this dialog from the page, 
so we have to validate first.