CMS7-7334: submit editor form when closing the editor via close link or context menu
All editor tabs are now in a form together too, so that the close-all/close-others/close-unmodified links will update all editor models. 

The validation mechanism interpreted a message as a message key and couldn't resolve it.
So now the violation contains the message model, putting the responsibility for coming up with a translation with the validator.