REPO-764 generalize hippolog item model to accommodate information other than workflow events
hippolog:category is now used for HippoEvent#category instead of for HippoWorkflowEvent#workflowCategory, the latter is now mapped to hippolog:workflowCategory. In fact all HippoEvent#values are mapped as is. Only the properties in the HippoEvent base class are part of the node type definition, all information belonging to extensions of HippoEvent (such as HippoWorkflowEvent) are accommodated by a residual property definition.