CMS7-7428: in IE, add stylesheets to the head using @import statements to work around the limit of 31 stylesheets
Revived Unico's patch for WICKET4254 by replacing Wicket.Head.Contributor.processLink at runtime with an IE-specific version.
The patch has been improved to properly handle CSS URLs that start with "./". Once added the URLs will become fully-qualified,
so the "./" should be excluded when checking whether new CSS URLs already exist.

CKEditor does not use the IE-workaround because it breaks the "styles" combobox (which is rendered in an iframe). 
The Wicket patch ensures that all CSS added in Ajax calls from the CMS do not pollute the head, so there's enough
room left for CKEditor to add it's own CSS styles.