REPO-838: make API friendlier for clients to use
The API forced a peculiar way of working with the returned objects.  Simplified the API by abstracting the instantiation & cleanup of the visitors. 
AFAICS, the ClassCastException in the Updater Info would never be thrown; use a class assignment check to validate whether the class is indeed a node updater visitor.