CMS7-7546: remove an image when it no longer contains a 'data-uuid' attribute
The old method of checking whether an image should be removed did not work when
additional attributes were set on it (e.g. an alignment, or a variant name) because
the 'remove' button in the image picker dialog does not set these attributes to blank. 
Just checking for the presence of the 'data-uuid' attribute is simpler and less brittle.