CMS7-7593: remove content element from dom when it's removed from the wicket component hierarchy
The client-side closing of the window happens in a timer.  The 'enter' triggered an onchange event on the form name was queued to the ajax channel, with the timer expiring during the creation of the new document.  This lead to the ajax request of the 'enter' being fired immediately, before the dialog window timer got the chance to clean up the DOM.
 
 Now explicitly removing the content from the DOM, synchronously when the response is processed.