CMS7-7632: revert creation of locale-specific names for documents and folders
The 'pretty name' of a new document or folder should be locale-independent. When an existing document or folder is renamed, the locale-specific names will be removed via DefaultWorkFlow#replaceAllLocalizedNames. When locale-specific names other than the one shown to the user exist, a warning message is shown in the rename dialog.