CMS7-7671: add 'Test Edit Menu' button that shows separate web app in an iframe
The IFramePanel is no longer a singleton and can be reused to create multiple iframes. The iframe used by the page editor is now retrieved via Ext.getCmp('pageEditorIFrame') instead of a globally known object.

The dummy app shown in the popup requires resources below '/angular' to be served by a ResourceServlet defined in the web.xml file of the cms application of an end project.