CMS7-7676: convert hst:cmseditmenu tags into 'edit menu' buttons
Processing of HTML comments in surfandedit.js is now split into one generic CommentsProcessor that iterates over all HTML comments, and two visitors (EditContentVisitor and EditMenuVisitor) that take care of converting a comment into a certain type of button.

The events 'showlinks' and 'hidelinks' have been renamed to 'show-edit-content-buttons' and 'hide-edit-content-buttons', respectively. Visibility of the 'edit menu' buttons is controlled by the events 'show-edit-menu-buttons' and 'hide-edit-menu-buttons'.