REPO-904: introduce HandleDocumentWorkflow
delegate to this workflow from the DocumentWorkflowImpl
the request on the DocumentHandle is now the request for the current user, assuming there is only one
version support has been removed from the document handle
the workflow manager now needs an additional 'delegating' property; when absent, the workflow will be instantiated on the variant for backwards compatibility.  When it is present (and true), the handle workflow will be returned.
a bit nicer methods on the WorkflowDefinition for the post-workflow-actions

the HandleDocumentWorkflow is currently very 'meta'.  Probably needs to become more explicit.