REPO-904: specific interface implementations
The reviewed actions interfaces have now been implemented in a more specific manner.
Instead of the 'delegating' property, use a 'subtype' on which to select a variant for the handle workflow.
Actually walk the class hierarchy when determining the Remote interfaces.
Few utility methods on the RepositoryTestCase to easier build up content.
Get editing & publication to pass.
import scxml definition as separate resource, using the xmlimport:file attribute