CMS7-7685: merge BasicReviewedActionsWorkflowPlugin, FullReviewedActionsWorkflowPlugin & UnlockWorkflowPlugin into the PreviewDocumentWorkflowPlugin
The translations for workflow actions have now also moved to the HandleDocumentWorkflowImpl.properties file.

Since I couldn't stand the sight of the big switch in the MenuHierarchy, I've made the submenu construction less magical.
ActionDescriptions can now provide a submenu key, that tells the system to put it in a particular sub-menu.

The WorkflowPlugin is now no longer needed; since there is only one workflow on a handle, this does not need to be enforced on the cms side.