HSTTWO-2833 Change parameter of JAX-RS move method
The frontend cannot send the uuid of the child before which a menu item should be
moved. It can send the index of where the moved item should be placed in the list
of child nodes. Therefore, the move method now requires this index as the last
path parameter. The node to be moved will be moved to the index specified.