CMS7-7696: move social sharing plugin into channel manager
The social sharing plugin makes more sense in the channel manager, as it needs a URL for it's operation.

Adjusted both socialsharing and channelactions to the new workflow UI configuration; they no longer need separate workflow categories with NullWorkflowImpl's, but are just part of the render cluster in the default workflow category.