HSTTWO-2852 Add lock info attributes to HTML comments of menu
The HstCmsEditMenuTag has been modified, so that it will return lock info
in HTML comments. I added the method HstRequestContext.getCmsUserID() to 
be able to compute the value for the "locked-by-current-user" boolean attribute.

To facilitate generation of (pseudo) JSON maps I have
created a TagUtils class.

TODO: rename the constants in ChannelManagerConstants