CMS7-7858: log out of fallbacksession at end of request
Since the fallbackSession does not contain any state, this change should have no adverse effects.
Logging out of the jcr session does reduce the load on the repository.

This change should reduce the memory & cpu impact of many requests to the login page.
E.g. as the result of a ping to check liveness.