CMS7-7905 Add mount path to page path
The mount path needs to be added to the page path to be able to open 
a page in the currently active channel. Furthermore, when creating or
editing a page, the mount path should be visible too. The page service
has been modified to fetch the mount info from the backend and the
mount path is used in the controllers to set the correct page host.