CMS7-7919: fix opening document specified by 'path' parameter
The listing was also expanded when there was already one open tab (for the document specified by the 'path' parameter). The logic is now simpler: only if there are no more open tabs, show the wide document listing (and reset the manual 'show narrow listing' override of the user) if the listing supports it.