CMS7-5805: move language.context.aware logic from NodePickerController to plugins that use the LinkPickerDialog
The previous commit gave the NodePickerController exact knowledge of the internal structure of the model passed to the NodePickerController, which did not seem right. With this commit, the plugins that instantiate a LinkPickerDialog use the LinkPickerDialogConfig helper object to determine the best 'base.uuid' parameter of the link picker dialog based on the locale of the document.
 
The new MockPluginTest base class makes it possible to unit test Wicket plugins (or any code that eventually accesses the UserSession) with mock nodes instead of a 'real' repository.