CMS7-7986: bump CKEditor version to 4.4.1-h1
Two new plugins are included and enabled by default:
- 'tableresize' make it possible to resize table columns by dragging their border
- 'textselection' retains the selected text and scroll position when switching between 
  normal view and source view

The new plugin 'codesnippet' is also included but not enabled by default. This plugin relies on 
the 'editor.addContentsCss' method (new in CKEditor 4.4) which is not available for editors in
a divarea (which Hippo CMS uses by default). We therefore add this method to each editor, 
and let it inject CSS into the HTML head of the CMS.

The auto-formatting implementation of the codemirror plugin also changed, and removes the 
indentation whitespace from code snippet blocks. Auto-formatting the HTML source is therefore 
not enabled by default anymore.