CMS7-8039 Add cache busting to resources fetched by AngularJS apps
I leveraged the server side generated antiCache parameter, so that I can use it for
cache busting static resources loaded by AngularJS (i.e. html and json).

For the REST calls I added a timestamp so that the call will not be cached.
For POST requests this is not required, but it does not harm if it's in the request.