CMS7-8041 Improve error feedback in channel manager if a site is down
Moved the try/catch blocks into the for loops, so that the loop is not terminated if one of the
web-apps is down. Otherwise, if one web-app is down, the channel list perspective would not
display any channels at all. This was not the reported issue however. I fixed the error feedback
by adding the context path to the error message, so that is clear which site is down.