CMS7-8182: select a menu item when clicked
The updated angular-ui-tree version caused a subtle change in the menu item selection. Previously, an item was immediately selected when the mouse button was pressed. Now a menu item is only selected when the mouse button is released and the menu item was not dragged into a new position. As a result, menu items can now be dragged into another position without selecting them.