CHANNELMGR-60: use ui-bootstrap tooltips instead of plain bootstrap tooltips
The new directive 'tooltipShow' watches an expression, and uses the custom events 'show' and 'hide' to trigger the tooltip of the associated element. This is used to show tooltips for invalid form fields.

Removed the usage of isolated scope from both the autoFocus and illegalCharacters directive because it does not work together with the tooltips (got a $compile:multidir error).

Reverted the 'autoFocus' directive to its old implementation: just focus the associated element. The expression watching was only used in the 'add page' dialog, which was questionable anyway. The old auto-focus directive 'hippo.channel.menu.auto-focus' was still referenced in various places bit did not exist anymore. These usages have been changed back to the new autoFocus directive.