HSTTWO-3116: fix unit test by running integration tests separately with surefire plugin
The TestWebResourceValveTest registers a mocked version of the WebResourcesService, which worked fine when the test was run in isolation. Yet it failed when run together with other tests because some other tests bootstrap a repository and start the real WebResourcesService. Since there can only be one named SingletonService at the time the registration of the mocked WebResourceService would then fail. The fix was to split unit tests and integration tests and run them separately, so the setup of the integration tests no longer interferes with the unit test fixtures and mocks.